Add comm sync functions
Added functions for creating, sending and waiting for the sync message.
The message is created from 3 characters that I randomly chose from the
extended ASCII table. The receiving device will wait for the sync
message indefinitely, since no time limit was implemented. This is well
within the parameters set for this project, since it guarantees that
only the desired message will be displayed.