Fix sync message exchange
Corrected the order in wich all bits and messages are sent; it now goes
from the most to the least significative.
When waiting for the sync message, the process originaly used was
optimal for a sync message correctely received in the first few trials.
Sice that was not showing to be the case, the process was simplified
and is now optmized to having a lot of misses before the first hit.
All received phrases are now made sure to end in NULL ('/0').
Added serial messsages for debugging the processes of sending and
receiving phrases.