Use interrupts for better comm precision
Changed the whole algorithm for timing the communication pin state. The
new way, using the processor's internal timer, has shown itself to be
extremely more precise then Arduino's delay and delaymicros functions.
Unfortunately, there is no official library for enabling the internal
timers in a Due, so the code will not work on a Mega or UNO. Also, there
is no guarantee that I'm not messing up with Serial comm or other native
functions that use internal clocks.