Make communications work
Added an extra header for verification on each package sent. This is not
compatible with the standard, but it works perfectly with only 1% of
bit error. There is still a lot of optimizations to make, such as
replacing all the validation algorithms by one that uses 8B10B, but this
here will be the plan B.